requirements.txt

bit~=0.8.0
redis~=4.5.5
fastapi~=0.95.2
uvicorn~=0.22.0
pydantic~=1.10.8
PyJWT~=2.7.0
bcrypt~=4.0.1